refactor: extract IssueExtractionFailurePresenter.swift from IssueExtractionPresenters.swift
Splits `IssueExtractionFailurePresenter` @MainActor class (~18 lines)
out of IssueExtractionPresenters.swift into its own file. That file
now holds only the IssueExtractionStatusPresenter enum + the
IssueExtractionFailurePresenter+Attempt extension.

Byte-identical move of lines 70-87. Both files import Foundation.

IssueExtractionPresenters.swift: 87 → 68 lines. Tests: 667.
